requirements.txt
asgiref==3.8.1
asttokens==2.4.1
crispy-bootstrap5==2024.10
da-vinci==0.4.0
decorator==5.1.1
Django==5.1
django-ckeditor==6.7.1
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-js-asset==2.2.0
django-thumbnails==0.8.0
executing==2.1.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
wcwidth==0.2.13
